pyproject.toml
[project]
name = "agent-browse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "patchright>=1.58.0",
    "pillow>=12.1.0",
]
